requirements.txt
google-generativeai
torch
huggingface_hub
transformers
python-dotenv
Flask
PyPDF2